Move path utilities to a separate module
Path utility functions that do not use `std::fs` are moved to a new
`path.rs` module in order to leave only functions that use `std::fs`
inside `fs.rs` as that provides better separation of concerns.
